apps/api/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pyjwt
svix
pytest
pytest-asyncio
httpx
aiosqlite
alembic
asyncpg
psycopg2-binary
greenlet
dnspython>=2.0
openai
google-genai
pgvector
